+++ components/assessments/InfoCard.vue
@@ -2,7 +2,7 @@
   <ApolloQuery
     v-slot="{ isLoading, result: { error } }"
     :query="require('~/gql/cards/getAssessmentsInfo.gql')"
-    :variables="{ courseCode, teacherView }"
+    :variables="{ courseCode, hideTakesStatus, learnerUsername, teacherView }"
     @result="setResult"
   >
     <generic-info-card
@@ -12,7 +12,14 @@
       :title="$tc('assessment._', 2)"
     >
       <template #default>
-        <stats-list v-if="!error" entity="assessment" :items="stats" />
+        <template v-if="!error">
+          <stats-list entity="assessment" :items="stats" />
+          <assessments-takes-status
+            v-if="!hideTakesStatus"
+            :assessments="assessments"
+            class="mx-2 mt-1"
+          />
+        </template>
         <span v-else>{{ $t('error.unexpected') }}</span>
       </template>
 
@@ -42,6 +49,14 @@ export default {
       type: Boolean,
       default: false,
     },
+    hideTakesStatus: {
+      type: Boolean,
+      default: false,
+    },
+    learnerUsername: {
+      type: String,
+      default: null,
+    },
     space: {
       type: String,
       required: true,
@@ -68,17 +83,24 @@ export default {
       return this.assessments?.length > 0
     },
     link() {
+      const learner =
+        this.teacherView && this.learnerUsername ? '-learners-username' : ''
+
       return {
         icon: 'mdi-view-list',
         text: this.$t('general.list'),
         to: {
-          name: `${this.space}-courses-code-assessments`,
-          params: { code: this.courseCode },
+          name: `${this.space}-courses-code${learner}-assessments`,
+          params: { code: this.courseCode, username: this.learnerUsername },
         },
       }
     },
     showCreateButton() {
-      return !this.hideCreateButton && this.course?.isCoordinator
+      return (
+        !this.hideCreateButton &&
+        this.course?.isCoordinator &&
+        !this.learnerUsername
+      )
     },
     stats() {
       if (!this.hasAssessments) {