+++ CHANGELOG.md
@@ -1,12 +1,33 @@
+## 1.2.3 (2023-03-04)
+
+### Features
+
+- **assessments:** display the assessments list and details for a given student (teacher)
+- **assessments:** display the general assessments takes status on the assessment info card (teacher and student)
+- **assessments:** display the takes status for each assessment on the assessments list and on the assessment detail page (teacher and student)
+
+### Improvements
+
+- added a progress bar type of information to the info panel component
+- defined a new assessment takes status info panel
+
+### Code Refactoring
+
+- refactored assessments list component
+
 ## 1.2.2 (2023-02-08)
 
 ### Features
+
+- **assessments:** configure an assessment to be a 'group' assessment or not (teacher)
 
 - **evaluations:** provide an optional acceptance comment when accepting an evaluation request (teacher)
 
 ### Bug Fixes
 
+- added the course publish button in the actions menu (teacher)
 - fixed the details page of learners by directly showing the competencies progress (teacher)
+- removed the invitation send button for unpublished courses (teacher)
 
 ## 1.2.1 (2023-02-04)